Re-add support for ShippingInstruction level carrierBookingReference
This is solved by:

 * On input, we always push down the SI level carrierBookingReference
   to the CargoItemTOs.  This way, the underlying code always work
   on the basis that CargoItemTOs /can/ have unique
   carrierBookingReferences without needing special cases for the
   SI-level variant.

 * On output, we attempt to "hoist" up the carrierBookingReference
   to the SI level as long as there is /exactly one/
   carrierBookingReference available.

Note that we are deliberately quite strict in not allowing "ambiguous"
carrierBookingReferences.  Either they are entirely on the SI level or
entirely on the CargoItemTO level.  Likewise on mapping the output, we
ensure that no CargoItemTO is "missing" a carrierBookingReference that
could affect whether we would hoist up the reference to the SI level.

Closes: #46
Signed-off-by: Niels Thykier <nt@asseco.dk>