Support but discard DisplayedAddress and PartyContactDetails
With these changes, EBL will now accept but ignore the
DisplayedAddress and PartyContactDetails (both via REST api endpoints
and the relevant SQL changes in the database).  That will come in a
future commit.

Currently, all work with these entities have been isolated in the
DocumentPartyService and for now they do not have their own service.

Signed-off-by: Niels Thykier <nt@asseco.dk>