DDT-965: Support admendmentToTransportDocument on SIs
Amendments to TDs that are `ISSU` are done by posting a new SI with
the desired content and a TD reference in the
`admendmentToTransportDocument` field. For now, the state flow for
these SIs are the same as regular SIs at least until the `DRFT` state
(and probably also the `APPR` state except charges might occur but we
do not do these).

To keep things simple, we do not allow the
`admendmentToTransportDocument` field to be changed in PUT.

Signed-off-by: Niels Thykier <nt@asseco.dk>