Remove test written on an invalid assumption
The test is built around `ShipmentEventService` returning an empty
Mono on error.  However, it never does that, which makes the
`.switchIfEmpty` and related mocking weird at best.  In theory, some
of the previous steps might trip an empty Mono but that will be fixed
in a later commit anyway (in a different way) and so the test is still
not useful there. The fix here will appear in the DDT-953 coming in a
separate PR.

With this analysis, remove the test and the related `switchIfEmpty`.

Signed-off-by: Niels Thykier <nt@asseco.dk>